+++ backend/canisters/escrow/impl/src/model/pending_payments_queue.rs
@@ -1,3 +1,4 @@
+use crate::model::swaps::Swap;
 use serde::{Deserialize, Serialize};
 use std::collections::VecDeque;
 use types::{TimestampMillis, TokenInfo, UserId};
@@ -10,6 +11,31 @@ pub struct PendingPaymentsQueue {
 impl PendingPaymentsQueue {
     pub fn push(&mut self, pending_payment: PendingPayment) {
         self.pending_payments.push_back(pending_payment);
+    }
+
+    pub fn push_refunds(&mut self, swap: &Swap, now: TimestampMillis) {
+        if swap.token0_received {
+            self.push(PendingPayment {
+                user_id: swap.created_by,
+                timestamp: now,
+                token_info: swap.token0.clone(),
+                amount: swap.amount0,
+                swap_id: swap.id,
+                reason: PendingPaymentReason::Refund,
+            });
+        }
+        if swap.token1_received {
+            if let Some((accepted_by, _)) = swap.accepted_by {
+                self.push(PendingPayment {
+                    user_id: accepted_by,
+                    timestamp: now,
+                    token_info: swap.token1.clone(),
+                    amount: swap.amount1,
+                    swap_id: swap.id,
+                    reason: PendingPaymentReason::Refund,
+                });
+            }
+        }
     }
 
     pub fn pop(&mut self) -> Option<PendingPayment> {